Change spelling mistakes and type of variable
Change spelling of attendantVerifiedBag to attendantVerifyBag and change type of begin variable.

+++ SCS - Software/src/org/lsmr/selfcheckout/customer/BaggingAreaController.java
@@ -20,7 +20,7 @@ public class BaggingAreaController extends TouchScreenController{
 	private ScanItemController scanItemControl;
 	private int numOfItemsInBaggingArea;
 	private double previousWeightOfCart;
-	private long begin;
+	private double begin;
 	private boolean askAttendantHelp = false;
